# Swagger_API_specification.yaml
openapi: 3.0.0
info:
  title: Account Management API
  description: APIs for managing customer accounts
  version: 1.0.0
servers:
  - url: /api
paths:
  /accounts:
    get:
      summary: Retrieve accounts by Customer ID
      description: Retrieves a list of accounts belonging to a customer by customer ID
      parameters:
        - name: customerId
          in: query
          required: true
          description: ID of the customer whose accounts to retrieve
          schema:
            type: integer
        - name: apiKey
          in: query
          required: true
          description: API key for authentication and rate limiting
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSummary'
        '401':
          description: Unauthorized
        '429':
          description: Too Many Requests
  /accounts/{accountNumber}:
    get:
      summary: Retrieve account details by account number
      description: Retrieves the details of an account by its account number
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: Account number of the account to retrieve
          schema:
            type: integer
        - name: apiKey
          in: query
          required: true
          description: API key for authentication and rate limiting
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
components:
  schemas:
    Account:
      type: object
      properties:
        accountNumber:
          type: integer
        accountName:
          type: string
        accountType:
          type: string
        customerId:
          type: integer
        bookBalance:
          type: number
        availableBalance:
          type: number
        blocked:
          type: boolean
        closed:
          type: boolean
        currency:
          type: string
        iban:
          type: string
        openingDate:
          type: string
          format: date
        lastModifiedDate:
          type: string
          format: date
    AccountSummary:
      type: object
      properties:
        accountNumber:
          type: integer
        accountName:
          type: string
        accountType:
          type: string
        currency:
          type: string